Convert to int

+++ lhc_web/modules/lhviews/loadinitialdata.php
@@ -8,7 +8,7 @@ erLhcoreClassChat::prefillGetAttributes($views, array('id', 'passive', 'name', '
 
 $response = [
     'views' => array_values($views),
-    'invites' => erLhAbstractModelSavedSearch::getCount(['limit' => false, 'filter' => ['status' => erLhAbstractModelSavedSearch::INVITE, 'user_id' =>  erLhcoreClassUser::instance()->getUserID()]])
+    'invites' => (int)erLhAbstractModelSavedSearch::getCount(['limit' => false, 'filter' => ['status' => erLhAbstractModelSavedSearch::INVITE, 'user_id' =>  erLhcoreClassUser::instance()->getUserID()]])
 ];
